requirements_server.txt
fastapi>=0.110
uvicorn>=0.27
PyJWT>=2.8
passlib[bcrypt]>=1.7
pydantic>=2.6